Refactor : 필터 드롭다운 구조 변경
