+++ src/components/common/Dropdown/FilterDropdown/DistrictsTags.tsx
@@ -1,16 +1,11 @@
-import { useState } from 'react';
 import XIcon from '../../../../assets/icons/XIcon';
 
 interface DistrictsTagsProps {
-  selectedDistrict: string[];
+  selectedDistricts: string[];
+  onDeleteTag: (removingDistrict: string) => void;
 }
-export default function DistrictsTags({ selectedDistrict }: DistrictsTagsProps) {
-  const [selectedDistricts, setSelectedDistricts] = useState(selectedDistrict);
 
-  const deleteDistrictTag = (removingDistrict: string) => {
-    const filteredDistricts = selectedDistricts.filter((district) => district !== removingDistrict);
-    setSelectedDistricts(filteredDistricts);
-  };
+export default function DistrictsTags({ selectedDistricts, onDeleteTag }: DistrictsTagsProps) {
   return (
     <div className="flex w-full flex-wrap gap-2">
       {selectedDistricts.map((district) => {
@@ -19,7 +14,7 @@ export default function DistrictsTags({ selectedDistrict }: DistrictsTagsProps) 
             district={district}
             key={district}
             onClick={() => {
-              deleteDistrictTag(district);
+              onDeleteTag(district);
             }}
           />
         );
@@ -30,11 +25,15 @@ export default function DistrictsTags({ selectedDistrict }: DistrictsTagsProps) 
 
 interface DistrictsTagProps {
   district: string;
-  onClick: () => void;
+  onClick: (district: string) => void;
 }
-function DistrictTag({ district }: DistrictsTagProps) {
+
+function DistrictTag({ district, onClick }: DistrictsTagProps) {
   return (
-    <div className="bg-red10 flex w-fit gap-1 rounded-[20px] px-2.5 py-1.5">
+    <div
+      className="bg-red10 flex w-fit gap-1 rounded-[20px] px-2.5 py-1.5"
+      onClick={() => onClick(district)}
+    >
       <p className="text-md-bold text-primary">{district}</p>
       <XIcon width="24" height="24" className="bg-black" />
     </div>
