+++ src/components/common/Dropdown/style.ts
@@ -1,5 +1,5 @@
 export const defaultContainerStyle =
-  'border-gray20 absolute rounded-md border cursor-pointer z-100';
+  'border-gray20 absolute rounded-md border-1 cursor-pointer z-100 ';
 
 export const defaultOptionStyle =
   'w-full hover:bg-red40 hover:text-white text-center py-3 hover:rounded-md ';
